requirements.txt
scipy
numpy
ipython
pillow
psutil
whiptail-dialogs
SQLAlchemy
